Fix last result in traces panel showing always last result for trace even for sub-agents
When a trace has multiple calls to sub-agents if you click on the
sub-agent the field for last result still shows last result for the full
traces. This was happening for all types. Prompt, Chat and External. We
centralized this logic into a hook that extract the information from the
children passed to the panel which can be the top span or any sub-span
like a sub-agent.
