+++ apps/web/src/components/tracing/spans/Chat.tsx
@@ -1,9 +1,5 @@
 import { MetadataItem } from '$/components/MetadataItem'
-import {
-  CompletionSpanMetadata,
-  SPAN_SPECIFICATIONS,
-  SpanType,
-} from '@latitude-data/constants'
+import { SPAN_SPECIFICATIONS, SpanType } from '@latitude-data/constants'
 import { Text } from '@latitude-data/web-ui/atoms/Text'
 import { ClickToCopy } from '@latitude-data/web-ui/molecules/ClickToCopy'
 import {
@@ -11,10 +7,7 @@ import {
   SPAN_COLORS,
   SpanFrontendSpecification,
 } from './shared'
-import { useTrace } from '$/stores/traces'
-import { findAllSpansOfType } from '@latitude-data/core/services/tracing/spans/fetching/findAllSpansOfType'
-import { findLastSpanOfType } from '@latitude-data/core/services/tracing/spans/fetching/findLastSpanOfType'
-import { useMemo } from 'react'
+import { useSpanCompletionData } from './useSpanCompletionData'
 import { formatCostInMillicents } from '$/app/_lib/formatUtils'
 import { MessageList } from '$/components/ChatWrapper'
 
@@ -27,60 +20,10 @@ export default {
 } satisfies SpanFrontendSpecification<SpanType.Chat>
 
 function DetailsPanel({ span }: DetailsPanelProps<SpanType.Chat>) {
-  const { data: trace } = useTrace({ traceId: span.traceId })
-  const completionSpan = useMemo(() => {
-    if (!trace) return undefined
-
-    return findLastSpanOfType({
-      children: trace.children,
-      spanType: SpanType.Completion,
-    })
-  }, [trace])
-  const completionSpans = useMemo(
-    () => findAllSpansOfType(trace?.children ?? [], SpanType.Completion),
-    [trace],
-  )
-  const aggregatedMetadata = useMemo(() => {
-    if (!completionSpans.length) return undefined
-
-    type AggregatedData = {
-      finishReason?: CompletionSpanMetadata['finishReason']
-      cost: number
-      tokens: {
-        prompt: number
-        cached: number
-        reasoning: number
-        completion: number
-      }
-    }
-
-    return completionSpans.reduce<AggregatedData>(
-      (acc, span) => {
-        const metadata = span.metadata as CompletionSpanMetadata | undefined
-        if (!metadata) return acc
-
-        return {
-          finishReason: metadata.finishReason,
-          cost: acc.cost + (metadata.cost || 0),
-          tokens: {
-            prompt: acc.tokens.prompt + (metadata.tokens?.prompt || 0),
-            cached: acc.tokens.cached + (metadata.tokens?.cached || 0),
-            reasoning: acc.tokens.reasoning + (metadata.tokens?.reasoning || 0),
-            completion:
-              acc.tokens.completion + (metadata.tokens?.completion || 0),
-          },
-        }
-      },
-      {
-        finishReason: undefined,
-        cost: 0,
-        tokens: { prompt: 0, cached: 0, reasoning: 0, completion: 0 },
-      },
-    )
-  }, [completionSpans])
-  const completionSpanMetadata = completionSpan?.metadata as
-    | CompletionSpanMetadata
-    | undefined
+  const { aggregatedMetadata, completionSpanMetadata } = useSpanCompletionData({
+    traceId: span.traceId,
+    spanId: span.id,
+  })
 
   return (
     <>
